Renaming MostViewedGrid and MostViewedList as subcomponents of MostViewed; reusing mocked data
