+++ cypress/lib/apiRoutes.js
@@ -10,10 +10,10 @@ export const setupApiRoutes = () => {
         method: 'GET',
         url: '/most-read/**',
     }).as('getMostRead');
-    // Listen for most-read call
+    // Listen for most-read-geo call
     cy.route({
         method: 'GET',
-        url: '/most-read-geo.json',
+        url: '**/most-read-geo**',
     }).as('getMostReadGeo');
     // Listen for most-read call
     cy.route({
