+++ src/web/components/MostViewed/MostViewedFooter/MostViewedFooter.mocks.tsx
@@ -1,0 +1,9 @@
+import { mockTab1, mockTab2 } from '../MostViewed.mocks';
+
+export const responseWithTwoTabs = {
+    data: [mockTab1, mockTab2],
+};
+
+export const responseWithOneTab = {
+    data: [mockTab1],
+};
