+++ src/web/components/MostViewed/MostViewedRight/MostViewedRight.mocks.tsx
@@ -1,0 +1,5 @@
+import { mockTab1 } from '../MostViewed.mocks';
+
+export const response = {
+    data: mockTab1,
+};
