--- src/web/components/MostViewedList/MostViewedList.mocks.tsx
@@ -1,127 +1,0 @@
-export const response = {
-    data: {
-        heading: 'Section header that should not appear',
-        trails: [
-            {
-                url:
-                    'https://www.theguardian.com/politics/2019/sep/15/eu-officials-reject-boris-johnson-claim-huge-progress-brexit-talks',
-                linkText:
-                    "LINKTEXT EU officials reject Boris Johnson claim of 'huge progress' in Brexit talks",
-                showByline: false,
-                byline: 'Jennifer Rankin and Daniel Boffey',
-                image:
-                    'https://i.guim.co.uk/img/media/85377038aacd71b2c0e55b0a55478165fe6d3014/0_0_6000_3600/master/6000.jpg?width=300&quality=85&auto=format&fit=max&s=de2cecf52b21492f01daa4520c4f2a97',
-                isLiveBlog: true,
-                pillar: 'news',
-            },
-            {
-                url:
-                    'https://www.theguardian.com/us-news/2019/sep/15/brett-kavanaugh-donald-trump-impeachment-supreme-court-justice',
-                linkText:
-                    'LINKTEXT Trump blasts calls for impeachment of Brett Kavanaugh after new allegations',
-                showByline: false,
-                byline: 'Martin Pengelly',
-                image:
-                    'https://i.guim.co.uk/img/media/579fd19481e46b9d6ed3c69c2a6992483df84478/0_0_6000_3600/master/6000.jpg?width=300&quality=85&auto=format&fit=max&s=63a6a5c13a8087e607a44ce5c7da2090',
-                isLiveBlog: false,
-                pillar: 'news',
-            },
-            {
-                url:
-                    'https://www.theguardian.com/sport/2019/sep/15/ashes-cricket-series-drawn-england-beat-australia-fifth-test',
-                linkText:
-                    'LINKTEXT England win fifth Test to draw Ashes series but Australia keep urn',
-                showByline: false,
-                byline: 'Vic Marks at the Oval',
-                image:
-                    'https://i.guim.co.uk/img/media/0df1fc1a2b78aa1772b4b43c2e5ebfc7f43111a8/0_91_5445_3267/master/5445.jpg?width=300&quality=85&auto=format&fit=max&s=3da9f89b54b0126fec2677117e73487e',
-                isLiveBlog: false,
-                pillar: 'news',
-            },
-            {
-                url:
-                    'https://www.theguardian.com/uk-news/2019/sep/15/boris-johnson-bonkers-plan-for-15bn-pound-bridge-derided-by-engineers',
-                linkText:
-                    "LINKTEXT Johnson's 'bonkers' plan for £15bn bridge derided by engineers",
-                showByline: false,
-                byline: 'Simon Murphy',
-                image:
-                    'https://i.guim.co.uk/img/media/57a9a1a56c7d319a8f57a9c8767793ae0481db4c/0_95_4950_2970/master/4950.jpg?width=300&quality=85&auto=format&fit=max&s=77949a713212b69499d0dc6b5bdb2faf',
-                isLiveBlog: false,
-                pillar: 'news',
-            },
-            {
-                url:
-                    'https://www.theguardian.com/politics/2019/sep/15/five-things-we-learned-from-david-cameron-memoir-boris-johnson-michael-gove-referendum',
-                linkText:
-                    "LINKTEXT Five things we learned from David Cameron's memoirs",
-                howByline: false,
-                byline: 'Caroline Davies',
-                image:
-                    'https://i.guim.co.uk/img/media/e05185cce0a0c6953e666276caa98f6ea104c989/0_214_3600_2160/master/3600.jpg?width=300&quality=85&auto=format&fit=max&s=188a7a9ed42517b9462ae0788fe96a0a',
-                isLiveBlog: false,
-                pillar: 'news',
-            },
-            {
-                url:
-                    'https://www.theguardian.com/sport/live/2019/sep/15/ashes-2019-england-v-australia-fifth-test-day-four-live',
-                linkText:
-                    'LINKTEXT Ashes 2019: England win fifth Test by 135 runs as series is drawn – as it happened',
-                showByline: false,
-                byline: 'Tanya Aldred,  Geoff Lemon  & Jonathan Howcroft',
-                image:
-                    'https://i.guim.co.uk/img/media/5792a3b971914e21bd04cb13eed54159463ce90e/0_128_4991_2994/master/4991.jpg?width=300&quality=85&auto=format&fit=max&s=6d4eab73546172d576a1022c29bdbcd2',
-                isLiveBlog: false,
-                pillar: 'news',
-            },
-            {
-                url:
-                    'https://www.theguardian.com/lifeandstyle/2019/sep/15/how-six-weddings-in-one-year-made-me-love-being-single',
-                linkText:
-                    'LINKTEXT How six weddings in one year made me love being single',
-                showByline: false,
-                byline: 'Sarah Johnson',
-                image:
-                    'https://i.guim.co.uk/img/media/2e87f0ee5f1a6c292cfcfc5f6f99e9a54ba3cc05/0_139_4032_2419/master/4032.jpg?width=300&quality=85&auto=format&fit=max&s=e951b289eda050060201927e978d30ef',
-                isLiveBlog: false,
-                pillar: 'sport',
-            },
-            {
-                url:
-                    'https://www.theguardian.com/fashion/2019/sep/15/joani-johnson-model-fenty-career-began-at-65-ageing-interview',
-                linkText:
-                    'LINKTEXT JoAni Johnson: the sexagenarian model defying convention',
-                showByline: false,
-                byline: 'Aaron Hicklin',
-                image:
-                    'https://i.guim.co.uk/img/media/2d88f2493b6eff49b6b8d6dd35345f03fa2b4af2/0_721_3361_2017/master/3361.jpg?width=300&quality=85&auto=format&fit=max&s=8601a174c9dc594f75318da8e99932bf',
-                isLiveBlog: false,
-                pillar: 'news',
-            },
-            {
-                url:
-                    'https://www.theguardian.com/fashion/2019/sep/15/victoria-beckham-announces-launch-of-own-makeup-brand',
-                linkText:
-                    'LINKTEXT Victoria Beckham launches makeup range and targets wellness market',
-                showByline: false,
-                byline: 'Jess Cartner-Morley',
-                image:
-                    'https://i.guim.co.uk/img/media/bceaab691b06c3f3830e7794d24c6ada2b301fad/847_177_2750_1650/master/2750.jpg?width=300&quality=85&auto=format&fit=max&s=3af893de9c01e38a8635c83883ef17ef',
-                isLiveBlog: false,
-                pillar: 'news',
-            },
-            {
-                url:
-                    'https://www.theguardian.com/commentisfree/2019/sep/15/government-boris-johnson-incredible-hulk-sam-gyimah',
-                linkText:
-                    'LINKTEXT There’s nothing normal about this beast of a government | Matthew d’Ancona',
-                showByline: true,
-                byline: 'Matthew d’Ancona',
-                image:
-                    'https://i.guim.co.uk/img/media/bc4e5ed4448eafd6238656aefae98c832cacc148/32_13_1898_1139/master/1898.jpg?width=300&quality=85&auto=format&fit=max&s=a4b8c116c5e580482dfe727d126b57ae',
-                isLiveBlog: false,
-                pillar: 'news',
-            },
-        ],
-    },
-};
